% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC}
\alias{AIC}
\title{AIC}
\usage{
AIC(covmat, p, m, n)
}
\arguments{
\item{covmat}{covariance matrix of innovations}

\item{p}{order of autoregression}

\item{m}{dimension of the variable}

\item{n}{number of observations}
}
\description{
AIC
}
