% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.BD.R
\name{AICc.BD}
\alias{AICc.BD}
\title{AICc (Brockwell and Davis (1991))}
\usage{
AICc.BD(covmat, p, m, n)
}
\arguments{
\item{covmat}{covariance matrix of innovations}

\item{p}{order of autoregression}

\item{m}{dimension of the variable}

\item{n}{number of observations}
}
\description{
AICc (Brockwell and Davis (1991))
}
