% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{AICc}
\alias{AICc}
\title{AICc (Hurvich and Tsai (1991))}
\usage{
AICc(covmat, p, m, n)
}
\arguments{
\item{covmat}{covariance matrix of innovations}

\item{p}{order of autoregression}

\item{m}{dimension of the variable}

\item{n}{number of observations}
}
\description{
AICc (Hurvich and Tsai (1991))
}
