% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{dgp}
\alias{dgp}
\title{Data generating process described in Hurvich and Tsai (1993)}
\usage{
dgp(sample_size, variance_covariance_matrix, ar_coefficients)
}
\arguments{
\item{sample_size}{Desired sample size}

\item{variance_covariance_matrix}{covariance matrix for the error term}

\item{ar_coefficients}{AR coefficients, no constant term. If AR(p) with m variables, it should be a vertical stack of all coefficient matrices (thus the dimension is mp*m)}
}
\value{
Returns a single series with sample size = sample_size
}
\description{
Data generating process described in Hurvich and Tsai (1993)
}
